MAPREDUCE-769. Make findbugs and javac warnings to zero. Contributed by Amareshwari Sriramadasu.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/mapreduce/trunk@825465 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/mapred/org/apache/hadoop/mapred/TestTaskTrackerBlacklisting.java
@@ -174,7 +174,7 @@ public class TestTaskTrackerBlacklisting extends TestCase {
         healthStatus.setLastReported(status.getLastReported());
       }
       jobTracker.heartbeat(tts, false, initialContact, 
-                           false, (short) responseId);
+                           false, responseId);
     }
     responseId++;
   }
